move main menu to separate file
