+++ Makefile
@@ -127,8 +127,8 @@ override CFLAGS	  += $(_CFLAGS) $(DEFINES) $(INCLUDES) \
 
 ### The object files (add further files here):
 
-OBJS = $(PLUGIN).o audio.o buf2rgb.o codec_audio.o codec_video.o config.o drmbuffer.o drmdevice.o drmplane.o grab.o h264parser.o \
-	logger.o mediaplayer.o pes.o queue.o ringbuffer.o softhddevice.o softhdsetupmenu.o softhdosd.o threads.o videorender.o videostream.o
+OBJS = $(PLUGIN).o audio.o buf2rgb.o codec_audio.o codec_video.o config.o drmbuffer.o drmdevice.o drmplane.o grab.o h264parser.o logger.o \
+	mediaplayer.o pes.o queue.o ringbuffer.o softhddevice.o softhdmenu.o softhdsetupmenu.o softhdosd.o threads.o videorender.o videostream.o
 
 ifeq ($(GLES),1)
 OBJS += openglosd.o
